requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
black==25.1.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
fuzzywuzzy==0.18.0
google-auth==2.40.3
google-genai==1.22.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Levenshtein==0.27.1
matplotlib-inline==0.1.7
mutagen==1.47.0
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
playwright==1.52.0
pluggy==1.6.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
PySocks==1.7.1
pytest==8.4.0
pytest-timeout==2.4.0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
PyYAML==6.0.1
RapidFuzz==3.13.0
requests==2.32.4
rsa==4.9.1
ruff==0.12.1
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
tenacity==8.5.0
tomli==2.2.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
